+++ entry/settings.py
@@ -13,7 +13,7 @@ CONFIG = Settings({
     'AND_SEARCH_CHARACTER': '&',
     'OR_SEARCH_CHARACTER': '|',
     'ESCAPE_CHARACTERS': ['(', ')', '<', '"', '{', '[', '#', '~', '@', '+', '*', '.', '?'],
-    'ESCAPE_CHARACTERS_REFERRALS_ENTRY': ['$', '(', '^', '|', '[', '+', '*', '.', '?'],
+    'ESCAPE_CHARACTERS_REFERRALS_ENTRY': ['$', '(', '^', '|', '[', '+', '*', '.', '?', '\\'],
     'ESCAPE_CHARACTERS_ENTRY_LIST': ['$', '(', '^', '\\', '|', '[', '+', '*', '.', '?'],
     'TIME_FORMAT': '%Y-%m-%dT%H:%M:%S',
 })